function for the rules
created a function that takes in a lenght 3 array and returns the value according to the rules of rule_30.
also To avoid boundary problems i made the inital matrix a (nxn+2)-matrix. then the plan is to stip away the the first and last column at the end (so a nxn matrix will still be returnes)